Add support for ESP8266_NONOS_SDK v3

+++ lib/button.c
@@ -1,9 +1,8 @@
-#include <stdint.h>
 #include "button.h"
 
 /* Initialises a button
  */
-void Button_init(struct Button *button, uint32_t pin, Button_state_reader state_reader, Button_time_reader time_reader)
+void Button_init(struct Button *button, unsigned int pin, Button_state_reader state_reader, Button_time_reader time_reader)
 {
     button->timestamp = 0;
     button->up = 0;
@@ -19,7 +18,7 @@ void Button_init(struct Button *button, uint32_t pin, Button_state_reader state_
  */
 void Button_read(struct Button *button)
 {
-    button->bounces = (button->bounces << 1) | (uint16_t)button->state_reader(button->pin);
+    button->bounces = (button->bounces << 1) | (unsigned int)button->state_reader(button->pin);
     button->up = button->down && (button->bounces > 0xFF00);
     button->longpress = button->down && button->timestamp && (button->time_reader() - button->timestamp > 3000000);
     if (button->down) {
